test: Fix failing DRM test on Tizen
The "basic flow" DRM test did not accurately simulate appending media
segments.  Instead of a start and end time, it used null, which is
only used in a full stack for init segments.  This triggered the
ContentWorkarounds module on a non-init segment in the context of this
test, which led to a failure.

The fix is simply to use a start and end time when appending the media
segment in the test.

Change-Id: If42651b5088ecf07fb33b57c5e86d4d6a56db162