test: Run tests on a custom Selenium grid
Includes a YAML config file for our lab, and documentation on how to
create a custom config for another Selenium grid.

Tests will be run nightly in the Shaka lab, using a self-hosted runner
with access to our private grid.

Tests can also be triggered manually by maintainers to test a PR in
the lab.  This will report status back to the PR.