ci: Re-run PR tests when leaving draft mode
Because our release PRs are created with the default GitHub token,
they do not trigger test runs.  This is a feature of GitHub that
prevents recursive loops of workflows.

This change makes it so that marking a PR "ready for review" will
trigger the tests to run again.  This gives us an easy way to trigger
test runs on a release PR, by marking it a draft, then marking it
"ready" again.