chore: Update AbortSignal polyfill for new compiler
This upgrades the compiler and reworks the AbortSignal polyfill to
match the new compiler externs for that class.  This is important to
make Shaka Player compatible with the latest compilers in use inside
Google.

Note that the Closure compiler is deprecated, so this should be our
final upgrade.  We will some day move to TypeScript.