chore: Raise screenshot threshold for native text
The screenshots with native text rendering don't always trigger our
threshold even when they should.  That's because the native text tends
to be much smaller.  Therefore the matching threshold should be
higher to make these smaller regions of text more sensitive to change.

This raises the threshold from 95% (still used for DOM-based) to 97%
for native text.  I believe this could have caught some of the changes
that I only caught later with manual review of screenshots.